chore(dts-generator): apply changes of other pull requests to reduce merge conflicts
