+++ test-packages/typed-json-model/webapp/model/test/cases/edgeCases.ts
@@ -67,61 +67,3 @@ const model4 = new TypedJSONModel(edgeCase);
 /** @expect ok     */ model4.setProperty("/anIntersection", { a: 1, b: "b" });
 /** @expect ts2345 */ model4.setProperty("/anIntersection", { a: 1 });
 /** @expect ts2322 */ model4.setProperty("/anIntersection", { a: 1, b: 2 });
-
-/***********************************************************************************************************************
- * Check model.bindContext
- **********************************************************************************************************************/
-
-// Absolute paths
-const data = {
-  aString: "string",
-  anObject: { a: "foo" },
-  anArray: [],
-  anArrayOfObjects: [{ aNumber: 1 }],
-  aPlaceholder: new Placeholder(),
-  anArrayOfPlaceholders: [new Placeholder()],
-  aTuple: ["string", 1],
-};
-
-const model5 = new TypedJSONModel(data);
-
-/** @expect ok     */ let clientContextBindingAbsolute: ClientContextBinding = model5.bindContext("/anObject");
-/** @expect ok     */ model5.bindContext("/anArrayOfObjects/0");
-/** @expect ok     */ model5.bindContext("/aPlaceholder");
-/** @expect ok     */ model5.bindContext("/anArrayOfPlaceholders/0");
-
-/** @expect ts2769 */ model5.bindContext("/anArray");
-/** @expect ts2769 */ model5.bindContext("/aTuple");
-/** @expect ts2769 */ model5.bindContext("/aTuple/0");
-/** @expect ts2769 */ model5.bindContext("/aJsonSafeArray/0");
-/** @expect ts2769 */ model5.bindContext("/anArrayOfObjects/0/aNumber");
-/** @expect ts2769 */ model5.bindContext("/anArray/0/doesNotExist");
-
-// Relative paths
-
-const dataForRelativeCase = {
-  root: {
-    aString: "string",
-    anObject: { a: "foo" },
-    anArray: [],
-    anArrayOfObjects: [{ aNumber: 1 }],
-    aPlaceholder: new Placeholder(),
-    anArrayOfPlaceholders: [new Placeholder()],
-    aTuple: ["string", 1],
-  },
-};
-
-const model6 = new TypedJSONModel(dataForRelativeCase);
-const context = model6.createBindingContext("/root");
-
-/** @expect ok     */ let clientContextBindingRelative: ClientContextBinding = model6.bindContext("anObject", context);
-/** @expect ok     */ model6.bindContext("anArrayOfObjects/0", context);
-/** @expect ok     */ model6.bindContext("aPlaceholder", context);
-/** @expect ok     */ model6.bindContext("anArrayOfPlaceholders/0", context);
-
-/** @expect ts2769 */ model6.bindContext("anArray", context);
-/** @expect ts2769 */ model6.bindContext("aTuple", context);
-/** @expect ts2769 */ model6.bindContext("aTuple/0", context);
-/** @expect ts2769 */ model6.bindContext("aJsonSafeArray/0", context);
-/** @expect ts2769 */ model6.bindContext("anArrayOfObjects/0/aNumber", context);
-/** @expect ts2769 */ model6.bindContext("anArray/0/doesNotExist", context);
